requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
async-generator==1.10
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
bitarray==2.9.2
blocknative-sdk==0.2.9
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.2
cytoolz==0.12.3
docutils==0.17.1
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
h11==0.12.0
hexbytes==0.3.1
idna==3.2
imagesize==1.4.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.5
myst-parser==0.18.0
outcome==1.1.0
packaging==24.0
parsimonious==0.10.0
protobuf==5.26.1
pycryptodome==3.20.0
Pygments==2.18.0
python-dotenv==1.0.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.1
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
toolz==0.12.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.11.0
urllib3==2.2.1
web3==6.18.0
websockets==12.0
wsproto==1.0.0
yarl==1.9.4
